+++ tests/base/attribute.cpp
@@ -4,16 +4,16 @@
 TEST(attribute, definition)
 {
 
-    float value = 1.0f;
+    double value = 1.0f;
     mv::Attribute attr(mv::AttrType::FloatType, value);
-    ASSERT_EQ(attr.getContent<float>(), value);
+    ASSERT_EQ(attr.getContent<double>(), value);
 
 }
 
 TEST(attribute, modification)
 {
-    float value = 1.0f, newValue = 2.0f;
+    double value = 1.0f, newValue = 2.0f;
     mv::Attribute attr(mv::AttrType::FloatType, value);
     attr.setContent(newValue);
-    ASSERT_EQ(attr.getContent<float>(), newValue);
+    ASSERT_EQ(attr.getContent<double>(), newValue);
 }
